Add trailets to select only the time calibraiton or the flat field calibration

+++ lstchain/scripts/onsite/onsite_create_calibration_file.py
@@ -29,6 +29,8 @@ optional.add_argument('-v', '--version', help="Version of the production",
 optional.add_argument('-s', '--statistics', help="Number of events for the flat-field and pedestal statistics",
                       type=int, default=10000)
 optional.add_argument('-b','--base_dir', help="Root dir for the output directory tree",type=str, default='/fefs/aswg/data/real')
+optional.add_argument('-t','--time_calibration', help="Perform only the time calibration (yes/no)",type=str, default='yes')
+optional.add_argument('-f','--ff_calibration', help="Perform only the charge calibration (yes/np)",type=str, default='yes')
 
 args = parser.parse_args()
 run = args.run_number
@@ -36,6 +38,8 @@ ped_run = args.pedestal_run
 prod_id = 'v%02d'%args.version
 stat_events = args.statistics
 base_dir = args.base_dir
+time_calibration = args.time_calibration
+ff_calibration = args.ff_calibration
 max_events = 1000000
 
 
@@ -74,14 +78,16 @@ def main():
         # produce ff calibration file
         #
 
+
         # define charge file names
         output_file = f"{output_dir}/calibration.Run{run}.0000.hdf5"
         log_file = f"{output_dir}/log/calibration.Run{run}.0000.log"
         print(f"\n--> Output file {output_file}")
-        if os.path.exists(output_file):
+        if os.path.exists(output_file) and ff_calibration is 'yes':
             if query_yes_no(">>> Output file exists already. Do you want to remove it?"):
                 os.remove(output_file)
             else:
+                print(f"\n--> Stop")
                 exit(1)
 
         print(f"\n--> Log file {log_file}")
@@ -93,31 +99,32 @@ def main():
             exit(1)
         print(f"\n--> Config file {config_file}")
 
-        # run lstchain script
-        cmd = f"lstchain_data_create_calibration_file " \
-              f"--input_file={input_file} --output_file={output_file} --pedestal_file={pedestal_file} " \
-              f"--FlatFieldCalculator.sample_size={stat_events} --PedestalCalculator.sample_size={stat_events}  " \
-              f"--EventSource.max_events={max_events} --config={config_file}  >  {log_file} 2>&1"
+        if ff_calibration is 'yes':
+            # run lstchain script
+            cmd = f"lstchain_data_create_calibration_file " \
+                  f"--input_file={input_file} --output_file={output_file} --pedestal_file={pedestal_file} " \
+                  f"--FlatFieldCalculator.sample_size={stat_events} --PedestalCalculator.sample_size={stat_events}  " \
+                  f"--EventSource.max_events={max_events} --config={config_file}  >  {log_file} 2>&1"
 
-        print("\n--> RUNNING...")
-        os.system(cmd)
+            print("\n--> RUNNING...")
+            os.system(cmd)
 
-        # plot and save some results
-        plot_file=f"{output_dir}/log/calibration.Run{run}.0000.pedestal.Run{ped_run}.0000.pdf"
-        print(f"\n--> PRODUCING PLOTS in {plot_file} ...")
-        calib.read_file(output_file)
-        calib.plot_all(calib.ped_data, calib.ff_data, calib.calib_data, run, plot_file)
+            # plot and save some results
+            plot_file=f"{output_dir}/log/calibration.Run{run}.0000.pedestal.Run{ped_run}.0000.pdf"
+            print(f"\n--> PRODUCING PLOTS in {plot_file} ...")
+            calib.read_file(output_file)
+            calib.plot_all(calib.ped_data, calib.ff_data, calib.calib_data, run, plot_file)
 
         #
         # produce drs4 time calibration file
         #
-
-        time_file = f"{output_dir}/time_calibration.Run{run}.0000.hdf5"
-        print(f"\n--> PRODUCING TIME CALIBRATION in {time_file} ...")
-        cmd = f"lstchain_data_create_time_calibration_file  --input_file {input_file} " \
-              f"--output_file {time_file} -conf {config_file} -ped {pedestal_file} 2>&1"
-        print("\n--> RUNNING...")
-        os.system(cmd)
+        if time_calibration is 'yes':
+            time_file = f"{output_dir}/time_calibration.Run{run}.0000.hdf5"
+            print(f"\n--> PRODUCING TIME CALIBRATION in {time_file} ...")
+            cmd = f"lstchain_data_create_time_calibration_file  --input_file {input_file} " \
+                  f"--output_file {time_file} -conf {config_file} -ped {pedestal_file} 2>&1"
+            print("\n--> RUNNING...")
+            os.system(cmd)
 
         print("\n--> END")
